Редактирует слайсы
