+++ project/src/store/offer-data/offer-data.ts
@@ -1,19 +1,53 @@
-import {createSlice} from '@reduxjs/toolkit';
+import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 import {NameSpace} from '../../components/const/const';
 import {OfferData} from '../../types/state';
 import {fetchOfferAction, fetchOfferByIdAction} from '../api-actions';
 
 const initialState: OfferData = {
   offers: [],
+  offer: null,
+  data: [],
+  selectPoint: 0,
+  sortType: 'Popular',
   isOffersDataLoading: false,
   hasError: false,
-  offer: null,
 };
+
+function getSortOffers(a:number, b:number) {
+  return (a - b);
+}
 
 export const offerData = createSlice({
   name: NameSpace.Data,
   initialState,
-  reducers: {},
+  reducers: {
+    changeOffer: (state, action: PayloadAction<{checkCity: string}>) => {
+      const {checkCity} = action.payload;
+      state.offers = state.data.filter((offer) => offer.city.name === checkCity);
+    },
+    selectPoint: (state, action: PayloadAction<{selectedPoint: number}>) => {
+      const {selectedPoint} = action.payload;
+      state.selectPoint = selectedPoint;
+    },
+    changeOption: (state, action: PayloadAction<{sortType: string}>) => {
+      const {sortType} = action.payload;
+
+      if (sortType === 'Price: low to high') {
+        const sortOffers = state.offers.sort((a, b) => getSortOffers(a.price, b.price));
+        state.offers = sortOffers;
+      } else if (sortType === 'Price: high to low') {
+        const sortOffers = state.offers.sort((a, b) => getSortOffers(b.price, a.price));
+        state.offers = sortOffers;
+      } else if (sortType === 'Top rated first') {
+        const sortOffers = state.offers.sort((a, b) => getSortOffers(b.rating, a.rating));
+        state.offers = sortOffers;
+      } else {
+        const sortOffers = state.offers.sort((a, b) => getSortOffers(a.id, b.id));
+        state.offers = sortOffers;
+      }
+      state.sortType = sortType;
+    },
+  },
   extraReducers(builder) {
     builder
       .addCase(fetchOfferAction.pending, (state) => {
@@ -21,6 +55,7 @@ export const offerData = createSlice({
         state.hasError = false;
       })
       .addCase(fetchOfferAction.fulfilled, (state, action) => {
+        state.data = action.payload;
         state.offers = action.payload;
         state.isOffersDataLoading = false;
       })
@@ -39,3 +74,5 @@ export const offerData = createSlice({
       });
   }
 });
+
+export const {changeOffer, changeOption, selectPoint} = offerData.actions;
