+++ project/src/store/review-process/review-process.ts
@@ -5,6 +5,7 @@ import {fetchReviewsAction, sendReviewAction} from '../api-actions';
 
 const initialState: ReviewProcess = {
   reviews: [],
+  hasError: false,
 };
 
 export const reviewProcess = createSlice({
@@ -16,15 +17,15 @@ export const reviewProcess = createSlice({
       .addCase(fetchReviewsAction.fulfilled, (state, action) => {
         state.reviews = action.payload;
       })
-      // .addCase(fetchReviewsAction.rejected, (state) => {
-      //   state.reviews = action.payload;
-      // })
-      // .addCase(fetchReviewsAction.fulfilled, (state, action) => {
-      //   state.reviews = action.payload;
-      // })
-      // .addCase(sendReviewAction.rejected, (state) => {
-      //   state.reviews = action.payload;
-      // })
+      .addCase(fetchReviewsAction.rejected, (state) => {
+        state.hasError = true;
+      })
+      .addCase(fetchReviewsAction.pending, (state, action) => {
+        state.hasError = false;
+      })
+      .addCase(sendReviewAction.rejected, (state) => {
+        state.hasError = false;
+      })
       .addCase(sendReviewAction.fulfilled, (state, action) => {
         state.reviews = action.payload;
       });
